restructured directory to use app routing
